Refactored the classes
